pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cliffcast"
version = "0.1.0"
description = "Transformer-based deep learning model for coastal cliff erosion prediction"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Connor J. Mack", email = "connorjmack@example.com"}
]
keywords = ["coastal erosion", "machine learning", "transformer", "lidar", "geomorphology"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
    # Core ML
    "torch>=2.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "einops>=0.7",

    # Data handling
    "pandas>=2.0",
    "xarray>=2023.1",
    "netCDF4>=1.6",
    "h5py>=3.8",

    # Point cloud
    "laspy>=2.5",

    # Geospatial
    "rasterio>=1.3",
    "geopandas>=0.14",
    "shapely>=2.0",
    "pyproj>=3.5",

    # Config & logging
    "pyyaml>=6.0",
    "omegaconf>=2.3",
    "wandb>=0.16",

    # Visualization
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "plotly>=5.18",

    # ML utilities
    "scikit-learn>=1.3",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.7",
    "jupyter>=1.0",
    "ipykernel>=6.25",
]

[project.urls]
Homepage = "https://github.com/connorjmack/transect-transformer"
Repository = "https://github.com/connorjmack/transect-transformer"
Issues = "https://github.com/connorjmack/transect-transformer/issues"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
src = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B905",  # zip() without an explicit strict= parameter
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
